+++ test/mneme_test.exs
@@ -1,6 +1,14 @@
 defmodule MnemeTest do
   use ExUnit.Case, async: true
 
+  doctest Mneme
+  doctest Mneme.Application
+  doctest Mneme.Collection
+  doctest Mneme.Error
+  doctest Mneme.Native
+  doctest Mneme.Pool
+  doctest Mneme.Result
+
   test "version/0 works" do
     assert Mneme.version() == "0.1.0"
   end